+++ pytsite/metatag/__init__.py
@@ -8,11 +8,11 @@ __license__ = 'MIT'
 
 
 def _init():
-    from pytsite import lang, events, tpl
+    from pytsite import lang, router, tpl
 
     lang.register_package(__name__)
-    events.listen('pytsite.router.dispatch', reset)
-    events.listen('pytsite.router.exception', lambda args: reset(args.get('title')))
+    router.on_dispatch(reset, -999)
+    router.on_exception(lambda args: reset(args.get('title')), -999)
 
     tpl.register_global('metatag', dump)
     tpl.register_global('metatags', dump_all)